Remove the list of supported printers from rastertoptch.c
Remove the list of supported printers from rastertoptch.c: the printers
actually supported are defined by the definitions in printer/; the duplicate
list in rastertoptch.c that summarizes some of the properties of those printers
serves as additional information in the literate programming documentation
generated by doxygen, which we're not really using anymore.  Also, keeping the
same information in two places just makes it harder to add new printers without
introducing inconsistencies.